requirements.txt
BeautifulSoup
Django==1.5.7
Jinja2
pillow
PyYAML
Pygments
South>=0.8.3
Sphinx
django-debug-toolbar==0.9.4
django-cache-panel
django-extensions
django-haystack<2.0
django-model-utils<1.5.0
django-tastypie
docutils
httplib2
lxml
mimeparse
pycrypto
pysolr==2.1.0
python-dateutil
ssh
uuid
wsgiref
simplejson
vectorformats
psycopg2

django-mptt

django-classy-tags
django-sekizai

django-tinymce
django-redis==3.3

django-recaptcha
django-disqus

django-grappelli==2.4.10
django-filebrowser==3.5.6

djangorestframework
markdown
numpy
scipy
pysal

werkzeug

requests==2.0.0
-e git://github.com/DeppSRL/django-widgets.git#egg=django-widgets

csvkit==0.7.3

